+++ frontend/my-adopthive-app/src/AdopteeHome/AdopteeHome.css
@@ -7,12 +7,11 @@
     align-items: center;
     justify-content: center;
     min-height: 100vh;
-    background-image: url(https://images.pexels.com/photos/9873934/pexels-photo-9873934.jpeg);
     background-size: cover;
     background-repeat: no-repeat;
-    color: white;
-
+    color: black;
   }
+  
 .home-content {
   text-align: center;
   margin-top: 30px;
