+++ frontend/my-adopthive-app/src/AdopteeList/AdopteeList.css
@@ -1,0 +1,59 @@
+.sidebar-container {
+    position: fixed;
+    top: 0;
+    left: 0;
+    width: 200px;
+    height: 100vh;
+    background-color: #333;
+    color: #fff;
+    transition: transform 0.5s ease-in-out;
+  }
+
+  .sidebar-container.open {
+    transform: translateX(0);
+  }
+
+  .sidebar-container.closed {
+    transform: translateX(-100%);
+  }
+
+  .sidebar-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    padding: 1rem;
+    background-color: #444;
+  }
+
+  .sidebar-header button {
+    font-size: 1.5rem;
+    cursor: pointer;
+  }
+
+  .sidebar-content {
+    padding: 1rem;
+  }
+
+  .sidebar-content ul {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+  }
+
+  .sidebar-content li {
+    padding: 1rem;
+    border-bottom: 1px solid #555;
+  }
+
+  .sidebar-content li:last-child {
+    border-bottom: none;
+  }
+
+  .sidebar-content a {
+    color: #fff;
+    text-decoration: none;
+  }
+
+  .sidebar-content a:hover {
+    color: #ccc;
+  }
