+++ frontend/my-adopthive-app/src/AdopteeProfile/AdopteeProfile.css
@@ -1,78 +1,75 @@
-.adoptees-profile-container {
-  width: 100%;
-  margin: auto;
-  padding: 20px;
-  background-color: black;
-  border: 1px solid #ccc;
-  color: white;
-  border-radius: 5px;
-  overflow-x: auto;
-}
-
-.adoptees-profile-container h2 {
-  text-align: center;
-}
-.adoptees-list {
-  list-style-type: none;
+body {
+  font-family: Arial, sans-serif;
+  margin: 0;
   padding: 0;
+  background-color: #f4f4f4;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 }
 
-.adoptees-item {
+form {
+  max-width: 800px;
+  margin: 150px auto;
+  padding: 20px;
+  background-color: #ffffff;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
+}
+
+form h3 {
   margin-bottom: 20px;
-  padding: 10px;
-  background-color: rgb(58, 21, 21);
-  border: 1px solid #e0e0e0;
-  color: white;
-  border-radius: 5px;
-  width: 90%;
-  margin: 20px;
-}
-
-.adoptees-item p {
-  margin: 5px 0;
   text-align: center;
+  font-size: 24px;
+  color: #333;
 }
 
-.adoptees-photo {
-  width: 150px;
-  height: auto;
-  border-radius: 5px;
-  display: block;
-  margin: 0 auto;
-
+input[type="text"],
+input[type="number"],
+input[type="date"],
+input[type="url"],
+select,
+textarea {
+  width: 97%;
+  padding: 10px;
+  margin: 10px 0;
+  border: 1px solid #ccc;
+  border-radius: 4px;
+  font-size: 16px;
 }
 
-.button-link-container {
-  margin-top: 20px;
+textarea {
+  height: 100px;
+  resize: vertical;
+}
 
+input::placeholder,
+textarea::placeholder {
+  color: #aaa;
 }
 
 button {
   padding: 10px 20px;
-  margin-top: 10px;
-  margin-bottom: 10px;
   font-size: 16px;
-  cursor: pointer;
-  background-color: rgb(58, 21, 21);
-  color: #fff;
+  color: #ffffff;
+  background-color: aquamarine;
   border: none;
-  border-radius: 5px;
-  margin-left: 30px;
-
+  border-radius: 4px;
+  cursor: pointer;
+  margin-top: 10px;
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
 }
 
 button:hover {
-  background-color: rgb(58, 21, 21);
+  background-color: aquamarine;
 }
 
-.view-matches-link {
-  padding: 10px 20px;
+button:disabled {
+  background-color: #ccc;
+  cursor: not-allowed;
+}
+
+.error {
+  color: red;
+  font-weight: bold;
+  margin-top: 10px;
   font-size: 16px;
-  text-decoration: none;
-  color:white;
-  margin-left: 30px;
-}
-
-.view-matches-link:hover {
-  text-decoration: underline;
 }
