+++ frontend/my-adopthive-app/src/AdopteeProfile/AdopteeProfile.jsx
@@ -1,6 +1,8 @@
 import { useState, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { UserContext } from '../UserContext';
+import './AdopteeProfile.css';
+import AdopteeSideBar from '../AdopteeSideBar/AdopteeSideBar';
 
 const AdopteeProfile = () => {
   const { user } = useContext(UserContext);
@@ -18,6 +20,7 @@ const AdopteeProfile = () => {
   const [city, setCity] = useState('');
   const [country, setCountry] = useState('');
   const [error, setError] = useState('');
+  const [isSideBarOpen, setIsSideBarOpen] = useState(false);
 
   const navigate = useNavigate();
 
@@ -54,27 +57,35 @@ const AdopteeProfile = () => {
     }
   };
 
+  function toggleSideBar()  {
+    setIsSideBarOpen(!isSideBarOpen);
+  }
+
   return (
-    <form onSubmit={handleSubmit}>
-      <input type="text" value={firstName} onChange={(e) => setFirstName(e.target.value)} placeholder="First Name" required />
-      <input type="text" value={lastName} onChange={(e) => setLastName(e.target.value)} placeholder="Last Name" required />
-      <input type="number" value={age} onChange={(e) => setAge(e.target.value)} placeholder="Age" required />
-      <select value={sex} onChange={(e) => setSex(e.target.value)} required>
-        <option value="Male">Male</option>
-        <option value="Female">Female</option>
-      </select>
-      <input type="date" value={birthdate} onChange={(e) => setBirthdate(e.target.value)} placeholder="Birthdate" required />
-      <input type="url" value={photoUrl} onChange={(e) => setPhotoUrl(e.target.value)} placeholder="Photo URL" />
-      <textarea value={background} onChange={(e) => setBackground(e.target.value)} placeholder="Background"></textarea>
-      <textarea value={interests} onChange={(e) => setInterests(e.target.value)} placeholder="Interests"></textarea>
-      <textarea value={education} onChange={(e) => setEducation(e.target.value)} placeholder="Education"></textarea>
-      <textarea value={traits} onChange={(e) => setTraits(e.target.value)} placeholder="Traits"></textarea>
-      <textarea value={dreams} onChange={(e) => setDreams(e.target.value)} placeholder="Dreams"></textarea>
-      <input type="text" value={city} onChange={(e) => setCity(e.target.value)} placeholder="City" required />
-      <input type="text" value={country} onChange={(e) => setCountry(e.target.value)} placeholder="Country" required />
-      {error && <div className="error">{error}</div>}
-      <button type="submit">Save Profile</button>
-    </form>
+    <div className='side-container'>
+      <AdopteeSideBar isOpen={isSideBarOpen} toggleSideBar={toggleSideBar} />
+      <form onSubmit={handleSubmit}>
+        <h3>Edit Profile</h3>
+        <input type="text" value={firstName} onChange={(e) => setFirstName(e.target.value)} placeholder="First Name" required />
+        <input type="text" value={lastName} onChange={(e) => setLastName(e.target.value)} placeholder="Last Name" required />
+        <input type="number" value={age} onChange={(e) => setAge(e.target.value)} placeholder="Age" required />
+        <select value={sex} onChange={(e) => setSex(e.target.value)} required>
+          <option value="Male">Male</option>
+          <option value="Female">Female</option>
+        </select>
+        <input type="date" value={birthdate} onChange={(e) => setBirthdate(e.target.value)} placeholder="Birthdate" required />
+        <input type="url" value={photoUrl} onChange={(e) => setPhotoUrl(e.target.value)} placeholder="Photo URL" />
+        <textarea value={background} onChange={(e) => setBackground(e.target.value)} placeholder="Background"></textarea>
+        <textarea value={interests} onChange={(e) => setInterests(e.target.value)} placeholder="Interests"></textarea>
+        <textarea value={education} onChange={(e) => setEducation(e.target.value)} placeholder="Education"></textarea>
+        <textarea value={traits} onChange={(e) => setTraits(e.target.value)} placeholder="Traits"></textarea>
+        <textarea value={dreams} onChange={(e) => setDreams(e.target.value)} placeholder="Dreams"></textarea>
+        <input type="text" value={city} onChange={(e) => setCity(e.target.value)} placeholder="City" required />
+        <input type="text" value={country} onChange={(e) => setCountry(e.target.value)} placeholder="Country" required />
+        { error && <div className="error">{error}</div>}
+        <button type="submit">Save Profile</button>
+      </form>
+    </div>
   );
 };
 export default AdopteeProfile;
