+++ frontend/my-adopthive-app/src/AdopteeSideBar/AdopteeSideBar.css
@@ -1,14 +1,17 @@
 .sidebar {
     position: fixed;
     left: 0;
-    top: 0;
+    top: 65px;
     display: block;
     bottom: 0;
-    width: 150px;
-    color: white;
+    width: 300px;
+    background-color: aquamarine;
+    color: black;
     overflow-y: auto;
+    height: calc(100vh - 60px);
     transition: transform 0.3s ease;
-    z-index: 1000;
+    z-index: 900;
+    box-shadow: 0 4px 8px black
 }
 
 .sidebar.open {
@@ -19,6 +22,11 @@
     padding: 20px;
 }
 
+.sidebar-items li i {
+    margin-right: 10px;
+    margin-right: 10px;
+}
+
 .sidebar ul {
     list-style-type: none;
     padding: 0;
@@ -27,6 +35,7 @@
 .sidebar ul li {
     margin-bottom: 100px;
     margin-top: 100px;
+
 }
 
 .sidebar ul li a {
@@ -40,7 +49,7 @@
 }
 
 .sidebar ul li a:hover {
-    background-color: rgb(58, 21, 21);
+    background-color: black;
 }
 
 .sidebar ul li a .fa-solid {
@@ -54,11 +63,11 @@
 }
 
 .sidebar ul li.selected a {
-    background-color: #007bff;
+    background-color: aquamarine;
 }
 
 .sidebar ul li.selected a:hover {
-    background-color: #0056b3;
+    background-color: aquamarine;
 }
 
 @media (max-width: 768px) {
