+++ frontend/my-adopthive-app/src/AdopterHome/AdopterHome.jsx
@@ -1,6 +1,5 @@
 import { useContext, useState, useEffect } from 'react';
 import { UserContext } from '../UserContext';
-import NavBar from '../NavBar/NavBar';
 import AdopterSideBar from '../AdopterSideBar/AdopterSideBar';
 import AdopteeListHome from '../AdopteeListHome/AdopteeListHome';
 import './AdopterHome.css';
@@ -28,8 +27,6 @@ function AdopterHome() {
     <div className="home-container">
       {user && (
         <div>
-          <NavBar isOpen={isSideBarOpen} toggleSideBar={toggleSideBar} />
-
           <AdopterSideBar isOpen={isSideBarOpen} toggleSideBar={toggleSideBar} />
           <div className="home-content">
             {isNewUser ? (
