+++ frontend/my-adopthive-app/src/AdopterProfile/AdopterProfile.jsx
@@ -1,6 +1,8 @@
 import { useState, useContext } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { UserContext } from '../UserContext';
+import Header from '../Header/Header';
+import AdopterSideBar from '../AdopterSideBar/AdopterSideBar';
 
 const AdopterProfile = () => {
   const { user } = useContext(UserContext);
@@ -15,6 +17,7 @@ const AdopterProfile = () => {
   const [country, setCountry] = useState('')
   const [error, setError] = useState('');
   const navigate = useNavigate();
+  const [isSideBarOpen, setIsSideBarOpen] = useState(false);
 
   const handleSubmit = async (e) => {
     e.preventDefault();
@@ -47,24 +50,31 @@ const AdopterProfile = () => {
       setError('Failed to save adopter profile')
     }
   };
+  function toggleSideBar()  {
+    setIsSideBarOpen(!isSideBarOpen);
+  }
 
   return (
-    <form onSubmit={handleSubmit}>
-      <input type="text" value={firstName} onChange={(e) => setFirstName(e.target.value)} placeholder="First Name" required />
-      <input type="text" value={lastName} onChange={(e) => setLastName(e.target.value)} placeholder="Last Name" required />
-      <input type="number" value={age} onChange={(e) => setAge(e.target.value)} placeholder="Age" required />
-      <select value={sex} onChange={(e) => setSex(e.target.value)} required>
-        <option value="Male">Male</option>
-        <option value="Female">Female</option>
-      </select>
-      <input type="text" value={status} onChange={(e) => setStatus(e.target.value)} placeholder="Status" required />
-      <input type="url" value={photoUrl} onChange={(e) => setPhotoUrl(e.target.value)} placeholder="Photo URL" />
-      <textarea value={background} onChange={(e) => setBackground(e.target.value)} placeholder="Background"></textarea>
-      <input type="text" value={city} onChange={(e) => setCity(e.target.value)} placeholder="City" required />
-      <input type="text" value={country} onChange={(e) => setCountry(e.target.value)} placeholder="Country" required />
-      {error && <div className="error">{error}</div>}
-      <button type="submit">Save Profile</button>
-    </form>
+    <div className='side-container'>
+      <Header/>
+      <AdopterSideBar isOpen={isSideBarOpen} toggleSideBar={toggleSideBar} />
+      <form onSubmit={handleSubmit}>
+        <input type="text" value={firstName} onChange={(e) => setFirstName(e.target.value)} placeholder="First Name" required />
+        <input type="text" value={lastName} onChange={(e) => setLastName(e.target.value)} placeholder="Last Name" required />
+        <input type="number" value={age} onChange={(e) => setAge(e.target.value)} placeholder="Age" required />
+        <select value={sex} onChange={(e) => setSex(e.target.value)} required>
+          <option value="Male">Male</option>
+          <option value="Female">Female</option>
+        </select>
+        <input type="text" value={status} onChange={(e) => setStatus(e.target.value)} placeholder="Status" required />
+        <input type="url" value={photoUrl} onChange={(e) => setPhotoUrl(e.target.value)} placeholder="Photo URL" />
+        <textarea value={background} onChange={(e) => setBackground(e.target.value)} placeholder="Background"></textarea>
+        <input type="text" value={city} onChange={(e) => setCity(e.target.value)} placeholder="City" required />
+        <input type="text" value={country} onChange={(e) => setCountry(e.target.value)} placeholder="Country" required />
+        {error && <div className="error">{error}</div>}
+        <button type="submit">Save Profile</button>
+      </form>
+    </div>
   );
 };
 export default AdopterProfile;
