+++ frontend/my-adopthive-app/src/Header/Header.css
@@ -1,16 +1,26 @@
 .header {
-    background-color: rgb(58, 21, 21);
-    width: 100%;
-    color: white;
-    text-align: left;
-    padding: 10px;
-    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.9);
-    overflow-x: hidden;
-    margin: 0px;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 10px 20px;
+  background-color: aquamarine;
+  color: black;
+}
 
-  }
+.header h1 {
+  margin: 0;
+}
 
-  .header h1 {
-    margin: 0;
-    font-size: 2em;
-  }
+.user-profile {
+  cursor: pointer;
+}
+
+.user-profile i {
+  font-size: 24px;
+}
+
+.profile-container {
+  position: absolute;
+  top: 70px; 
+  right: 20px;
+}
