+++ frontend/my-adopthive-app/src/Profile/Profile.css
@@ -1,56 +1,65 @@
 .profile {
-  background-color: rgb(58, 21, 21);
-  padding: 10px;
-  top: 70px;
-  width:25%;
-  right: 150px;
-  border-radius: 7px;
-  z-index: 1000;
-  box-shadow: 0 4px 8px black;
+  background-color: aquamarine;
+  color: #333;
+  padding: 20px;
+  width: 250px;
+  border-radius: 8px;
+  box-shadow: 0 4px 8px rgba(0, 0, 0, 0.2);
   position: absolute;
-  text-align: left;
+  top: 70px;
+  right: 60px;
+  text-align: center;
+  z-index: 1000;
 }
 
 .profile-image-placeholder {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100px;
-    height: 100px;
-    border-radius: 50%;
-    background-color: #000;
-    cursor: pointer;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  color: black;
+  justify-content: center;
+  width: 130px;
+  height: 130px;
+  border-radius: 50%;
+  background-color: #333;
+  cursor: pointer;
+  margin-bottom: 10px;
 }
 
+.profile-image-placeholder img,
 .profile img {
-    margin-bottom: 10px;
-    object-fit: cover;
-    border-radius: 50%;
-    margin-bottom: 10px;
-    width: 130px;
-    height: 130px;
+  width: 130px;
+  height: 130px;
+  border-radius: 50%;
+  object-fit: cover; 
 }
 
 .profile p {
-    margin-bottom: 5px;
-    margin-top: 5px;
+  margin: 5px 0;
+  font-size: 1em;
+  text-align: left;
 }
 
 .profile input {
-    margin-top: 5px;
+  margin-top: 10px;
+  display: block;
 }
 
 .profile button {
-    margin-top: 10px;
-    padding: 8px 16px;
-    background-color: black;
-    color: white;
-    border: none;
-    cursor: pointer;
-    margin-left: 30%;
+  margin-top: 15px;
+  padding: 10px 20px;
+  background-color: white;
+  border: none;
+  border-radius: 4px;
+  cursor: pointer;
+  font-size: 1em;
 }
 
 .profile button:hover {
-    background-color:black;
+  background-color: #333;
+}
+
+.profile button:disabled {
+  background-color: #666;
+  cursor: not-allowed;
 }
